mcp-token-auditor/requirements.txt
# MCP Token Auditor - Dependencies

# Core async framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# Token counting
tiktoken>=0.5.2

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0

# HTTP client
httpx>=0.26.0
aiohttp>=3.9.1

# Configuration
pyyaml>=6.0.1
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Utilities
uuid6>=2024.1.12
python-json-logger>=2.0.7

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0

# Type checking
mypy>=1.8.0
